move lib to src
